fix: project 'word monitoring' for theme 'jcf' (refactoring)
